Added favorite selection and deselection
I passed a listener (favoriteSelectionListener) to the Viewholder to detects clicks on the favorite ImageView. When the favorite ImageView is clicked it adds the collection as a favorite if it isn't and if it's already a favorite I delete it from the favorites DB.

I chose to switch the image resources to indicate immediately the favorite Image gets clicked and also change the isFavorite Variable on the Collection Object.

I chose the method because it is faster instead of fetching the list of all saved favorites eveytime the favorite ImageView get's clicked.