Indicating favorite collection selection
I added the ability to indicate the favorite Collections whenever the collection lists gets fetched.

I first fetched the list of IDs of the favorite collection items. While mapping, I then check the id of each collection item if it is contained in the list of the IDs of the favorite items.

This way, whenever the user comes back into the app, the favorite items will be indicated