Added favorites List
Added the implementation for showing the list of favorite collection as well as a toggle to switch between all collection and favorites.

I created a switch since the selection has only 2 options and I also reset the Switch back to 'All collections' when a user swipes down to refresh because swipeToRefresh fetches and shows all Collections.